+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Bugfixes
 
 * Fix exporting and importing content template.
 * Fix error when content matching to a text template is saved.
+* Fix duplicated content field values when updating content.
 * All commits in `0.10.0`_.
 
 Security